# Conflicts:
#	Backend/Backend/src/main/java/com/bankapp/Backend/config/SecurityConfig.java

+++ Backend/Backend/src/main/java/com/bankapp/Backend/model/Account.java
@@ -1,7 +1,6 @@
 package com.bankapp.Backend.model;
 
 import jakarta.persistence.*;
-import java.math.BigDecimal;
 
 @Entity
 @Table(name= "accounts")
@@ -20,24 +19,24 @@ public class Account
     private String iban;
 
     @Column(nullable = false)
-    private BigDecimal balance;
+    private double balance;
 
     @Column(name = "account_type", nullable=false)
     private String accountType; // e.g., "checking", "savings", "debit", "credit"
 
     @Column(name = "daily_tansfer_limit", nullable = false)
-    private BigDecimal dailyTransferLimit;
+    private double dailyTransferLimit;
 
     @Column(name = "absolute_transfer_limit", nullable = false)
-    private  BigDecimal absoluteTransferLimit;
+    private  double absoluteTransferLimit;
 
     @Column(nullable = false)
     private String status; // e.g., "active", "unapproved", "frozen"
 
     public Account() {}
 
-    public Account(User user, String iban, BigDecimal balance, String accountType,
-                   BigDecimal dailyTransferLimit, BigDecimal absoluteTransferLimit, String status) {
+    public Account(User user, String iban, double balance, String accountType,
+                   double dailyTransferLimit, double absoluteTransferLimit, String status) {
         this.user = user;
         this.iban = iban;
         this.balance = balance;
@@ -56,17 +55,17 @@ public class Account
     public String getIban() { return iban; }
     public void setIban(String iban) { this.iban = iban; }
 
-    public BigDecimal getBalance() { return balance; }
-    public void setBalance(BigDecimal balance) { this.balance = balance; }
+    public double getBalance() { return balance; }
+    public void setBalance(double balance) { this.balance = balance; }
 
     public String getAccountType() { return accountType; }
     public void setAccountType(String accountType) { this.accountType = accountType; }
 
-    public BigDecimal getDailyTransferLimit() { return dailyTransferLimit; }
-    public void setDailyTransferLimit(BigDecimal dailyTransferLimit) { this.dailyTransferLimit = dailyTransferLimit; }
+    public double getDailyTransferLimit() { return dailyTransferLimit; }
+    public void setDailyTransferLimit(double dailyTransferLimit) { this.dailyTransferLimit = dailyTransferLimit; }
 
-    public BigDecimal getAbsoluteTransferLimit() { return absoluteTransferLimit; }
-    public void setAbsoluteTransferLimit(BigDecimal absoluteTransferLimit) { this.absoluteTransferLimit = absoluteTransferLimit; }
+    public double getAbsoluteTransferLimit() { return absoluteTransferLimit; }
+    public void setAbsoluteTransferLimit(double absoluteTransferLimit) { this.absoluteTransferLimit = absoluteTransferLimit; }
 
     public String getStatus() { return status; }
     public void setStatus(String status) { this.status = status; }
